agent/requirements.txt
livekit-agents[google,images,turn-detector]~=1.2,>=1.2.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv
googlemaps>=4.10.0
